fixed box-sizing assumptions
The default box-sizing is content-box, not border-box. This causes a layout issue in IE 10 and probably others when no box-sizing is set and $.browser.boxModel is not available. Honestly, there is no reason to even make all those browser version checks anyway. Just fetch the boxSizing value and if its not available or not set, default to content-box.

See: https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing